Was not compiling.

+++ DependentExtensions/Lobby2/PGSQL/Lobby2Message_PGSQL.cpp
@@ -2688,7 +2688,7 @@ bool RakNet::Ranking_GetTotalScore_PGSQL::ServerDBImpl( Lobby2ServerCommand *com
 		return true;
 	}
 
-	int64_t count;
+	long long count;
 	PostgreSQLInterface::PQGetValueFromBinary(&count, result, 0, "count");
 	numScoresSubmitted = (unsigned int) count;
 	if (numScoresSubmitted>0)
